db/requirements.txt
fastapi==0.95.0
uvicorn==0.23.1
sqlalchemy==2.0.16
psycopg2==2.9.3
pydantic==1.10.2
python-dotenv==0.20.0
passlib==1.7.4
pyjwt==2.6.0
